cNetwork: Split the main cpp file into several files.
